% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4arena.R
\name{c4arena}
\alias{c4arena}
\title{Pit two players in a game of c4!}
\usage{
c4arena(player1, player2, nrow = 6, ncol = 7)
}
\arguments{
\item{player1}{Either an instance of a reference class or a reference class generator}

\item{player2}{Either an instance of a reference class or a reference class generator}

\item{nrow}{Number of rows for the board}

\item{ncol}{Number of columns for the board}
}
\value{
Final game state
}
\description{
Pit two players in a game of c4!
}
\examples{
c4arena(randomPlayer, randomPlayer)

p1 <- randomPlayer$new()
p2 <- randomPlayer$new()
c4arena(p1, p2)

\dontrun{
p1 <- randomPlayer$new(name = "Bob")
p2 <- human$new(name = "Alice")
c4arena(p1, p2)
}
}
