% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4game.R
\docType{class}
\name{c4game-class}
\alias{c4game-class}
\alias{c4game}
\title{A reference class to create a connect 4 game object.}
\value{
a reference class (RC/R5) object.
}
\description{
A reference class to create a connect 4 game object.
}
\section{Fields}{

\describe{
\item{\code{board}}{A matrix.}

\item{\code{y}}{A character.}

\item{\code{player}}{A number}
}}


\examples{
c4game$new(board = matrix(, nrow = 6, ncol = 7), player = 1)
returns an r5 object
}
