% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c4game.R
\docType{data}
\name{c4game}
\alias{c4game}
\title{A reference class to create a connect 4 game object.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
c4game
}
\value{
an R6 object.
}
\description{
A reference class to create a connect 4 game object.
}
\section{Fields}{

\describe{
\item{\code{board}}{A matrix.}

\item{\code{y}}{A character.}

\item{\code{player}}{A number}
}}

\examples{
game <- c4game$new()
game <- c4game$new(board = matrix(, nrow = 6, ncol = 7))
# object instatiation returns an R6 object

}
\keyword{datasets}
