% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player-human.R
\docType{class}
\name{human-class}
\alias{human-class}
\alias{human}
\title{Human player for c4}
\description{
Human player for c4
}
\section{Fields}{

\describe{
\item{\code{name}}{character.}
}}


\examples{
human$new(name="Mark")

}
