% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player-human.R
\name{human_make_move}
\alias{human_make_move}
\title{Make a human move}
\arguments{
\item{board}{matrix from game}

\item{lastMove}{the laste move made}
}
\value{
the player's move
}
\description{
Print the board and give a prompt
}
