% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player-human.R
\name{human_rejected}
\alias{human_rejected}
\title{Let the player know the move was rejected}
\arguments{
\item{board}{matrix from game}

\item{lastMove}{the laste move made}
}
\value{
the player's new move
}
\description{
Let the player know the move was rejected
}
