% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playRandomGame.R
\name{playRandomGame}
\alias{playRandomGame}
\title{Plays a random game of connect 4.}
\usage{
playRandomGame(verbose = FALSE)
}
\arguments{
\item{verbose}{= FALSE}
}
\description{
Plays a random game of connect 4.
}
\examples{
playRandomGame()
Player 2 Wins!
Moves: 25
}
