% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player-random.R
\docType{class}
\name{randomPlayer-class}
\alias{randomPlayer-class}
\alias{randomPlayer}
\title{Random player for c4}
\description{
Random player for c4
}
\section{Fields}{

\describe{
\item{\code{name}}{character.}
}}


\examples{
randomPlayer$new(name="Bot1")
}
