% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player-sequence.R
\docType{class}
\name{sequencePlayer-class}
\alias{sequencePlayer-class}
\alias{sequencePlayer}
\title{Random player for c4}
\description{
Give the bot the sequence and it will make the moves. Cyclic repetition means
it never runs out of ideas.
}
\section{Fields}{

\describe{
\item{\code{name}}{character. Name of the bot}

\item{\code{moveSeq}}{numeric. The sequence of moves. Should be column indexes}

\item{\code{moveCnt}}{numeric. Private. Where you are in the sequence}
}}


\examples{
sequencePlayer$new(moveSeq = 1:6, name="SequenceBot1")

}
